Time changed from float to double Fixed updating Time by sdl3

+++ src/Engine/src/Gui/ImGui/FpsCounter.h
@@ -15,7 +15,7 @@ namespace BeeEngine::Internal
         void Render() override;
         void OnEvent(EventDispatcher& event) override;
     private:
-        float m_LastTime = Time::TotalTime();
+        double m_LastTime = Time::TotalTime();
         float m_Fps = 0;
         float m_CurrentFps = 0;
     };
